+++ docs/changelog.md
@@ -3,6 +3,9 @@
 This project follows [Semantic Versioning](https://semver.org/) and [Keep A Change Log](https://keepachangelog.com/en/1.0.0/).
 
 ## [v0.3.0] WIP
+### Added
+- Built in reconnect support.
+- Support giving an event loop to DiscordClient.run()
 
 ## [v0.2.0] 2021-10-18
 ### Added